LayoutTests/http/tests/security/xssAuditor/malformed-HTML-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Ca%3Cimg/src/onerror=alert(1)//%3C' because its source code was found within the request. The auditor was enabled as the server sent neither an 'X-XSS-Protection' nor 'Content-Security-Policy' header.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
